fix maven build
