+++ code/src/graphics/gamefield/Edge.java
