+++ code/src/graphics/gamefield/JCistern.java
@@ -1,27 +1,30 @@
 package graphics.gamefield;
 
-import model.fields.WaterSource;
+import model.fields.Cistern;
 
 import javax.swing.*;
 import java.awt.*;
+import java.util.Objects;
 
 /**
- * WaterSourcet reprezentáló grafikus osztály
+ * Ciszternákat reprezentáló grafikus osztály
  */
-public class JWaterSource extends Vertex{
-    WaterSource observedWaterSource;
+public class JCistern extends Vertex {
+    Cistern observedCistern;
+
 
     /**
-     * Létrehozza magát a kapott helyere
-     * @param center középpontja
+     * Létrehozza magát paraméterrel
+     *
+     * @param center hova legyen kirajzolva
      */
-    public JWaterSource(Point center, WaterSource ws){
-        super(center);
-        observedWaterSource=ws;
-        color= Color.blue;
+    public JCistern(Point center, Cistern cist) {
+        this.center=center;
+        observedCistern = cist;
+        color = Color.orange;
 
         //Alap textura betöltése
-        ImageIcon tmpIcon = new ImageIcon("res/graphics/objects/watersource.png");
+        ImageIcon tmpIcon = new ImageIcon(Objects.requireNonNull(getClass().getResource("/graphics/objects/cistern.png")));
         Image resizedTmpImage = tmpIcon.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH);
         graphicalImage = new ImageIcon(resizedTmpImage);
     }
@@ -29,6 +32,7 @@ public class JWaterSource extends Vertex{
 
     /**
      * Kirajzolja magát
+     *
      * @param g the <code>Graphics</code> object to protect
      */
     @Override
