+++ code/src/graphics/gamefield/Vertex.java
