+++ code/src/model/CountableObserved.java
